Fix blocking write issue.
When primarily using just one of the terminals, a bug will manifest as the
terminal simply stops working. Opening the other terminal will unblock both.
This happens since the write hits the limit of how much it can write without it
being read out the other side, and the unused terminal is never read.

Just set each to NONBLOCKING to avoid.